Compat SOG - Remove hard `ace_hearing` & `ace_refuel` dependencies (#10655)
