+++ addons/compat_sog/compat_sog_hearing/CfgWeapons.hpp
@@ -1,0 +1,35 @@
+class CfgWeapons {
+    class vn_b_headgear_base;
+    class vn_b_helmet_aph6_01_01: vn_b_headgear_base {
+        HEARING_PROTECTION_VICCREW;
+    };
+    class vn_b_helmet_aph6_02_01: vn_b_headgear_base {
+        HEARING_PROTECTION_VICCREW;
+    };
+    class vn_b_helmet_sph4_01_01: vn_b_headgear_base {
+        HEARING_PROTECTION_VICCREW;
+    };
+    class vn_b_helmet_sph4_02_01: vn_b_headgear_base {
+        HEARING_PROTECTION_VICCREW;
+    };
+    class vn_b_helmet_t56_01_01: vn_b_headgear_base {
+        HEARING_PROTECTION_VICCREW;
+    };
+    class vn_b_helmet_svh4_01_01: vn_b_headgear_base {
+        HEARING_PROTECTION_VICCREW;
+    };
+    class vn_b_helmet_svh4_02_01: vn_b_headgear_base {
+        HEARING_PROTECTION_VICCREW;
+    };
+
+    class vn_o_headgear_base;
+    class vn_o_helmet_tsh3_01: vn_o_headgear_base {
+        HEARING_PROTECTION_VICCREW;
+    };
+    class vn_o_helmet_zsh3_01: vn_o_headgear_base {
+        HEARING_PROTECTION_VICCREW;
+    };
+    class vn_o_helmet_shl61_01: vn_o_headgear_base {
+        HEARING_PROTECTION_VICCREW;
+    };
+};
