+++ addons/compat_sog/compat_sog_hearing/config.cpp
@@ -1,5 +1,5 @@
 #include "script_component.hpp"
-#include "\z\ace\addons\refuel\defines.hpp"
+#include "\z\ace\addons\hearing\script_macros_hearingProtection.hpp"
 
 class CfgPatches {
     class SUBADDON {
@@ -7,15 +7,17 @@ class CfgPatches {
         units[] = {};
         weapons[] = {};
         requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = {"loadorder_f_vietnam", "ace_refuel"};
+        requiredAddons[] = {
+            "loadorder_f_vietnam",
+            "ace_hearing"
+        };
         skipWhenMissingDependencies = 1;
         author = ECSTRING(common,ACETeam);
-        authors[] = {"johnb43"};
         url = ECSTRING(main,URL);
         VERSION_CONFIG;
+
+        addonRootClass = QUOTE(ADDON);
     };
 };
 
-class CBA_Extended_EventHandlers;
-
-#include "CfgVehicles.hpp"
+#include "CfgWeapons.hpp"
