+++ addons/compat_sog/compat_sog_hearing/script_component.hpp
@@ -1,0 +1,3 @@
+#define SUBCOMPONENT hearing
+#define SUBCOMPONENT_BEAUTIFIED Hearing
+#include "..\script_component.hpp"
