+++ addons/compat_sog/config.cpp
@@ -1,6 +1,4 @@
 #include "script_component.hpp"
-#include "\z\ace\addons\refuel\defines.hpp"
-#include "\z\ace\addons\hearing\script_macros_hearingProtection.hpp"
 
 class CfgPatches {
     class ADDON {
